Add playwright scripts to obtain platform tokens and admin auth cookies and remove from test
